+++ packages/vsce/__tests__/__e2e__/specs/program.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "@playwright/test";
-import { constants, findAndClickText, findAndClickTreeItem, getTreeItem, prepareZoweExplorerView, resetWiremock, resetZoweExplorerView } from "../utils/helpers";
+import { constants, findAndClickText, findAndClickTreeItem, getClipboardContent, getTreeItem, prepareZoweExplorerView, resetWiremock, resetZoweExplorerView } from "../utils/helpers";
 
 test.beforeEach(async ({ page, request }) => {
   await resetWiremock(request);
@@ -83,5 +83,24 @@ test.describe("Program tests", () => {
 
     await expect(getTreeItem(page, `${constants.LIBRARY_1_NAME} (DSNAME='MYLIBDS1') AND (LIBRARY='${constants.LIBRARY_1_NAME}')`, false)).toHaveCount(1);
     await expect(getTreeItem(page, constants.LIBRARY_DS_1_NAME)).toHaveCount(1);
+  });
+
+  test("should copy program name", async ({ page }) => {
+    await findAndClickTreeItem(page, constants.PROFILE_NAME);
+    await findAndClickTreeItem(page, constants.CICSPLEX_NAME);
+    await findAndClickTreeItem(page, constants.REGION_NAME);
+    await findAndClickTreeItem(page, "Programs");
+
+    await expect(getTreeItem(page, constants.PROGRAM_1_NAME)).toBeVisible();
+    await expect(getTreeItem(page, constants.PROGRAM_1_NAME)).toHaveText(constants.PROGRAM_1_NAME);
+
+    await findAndClickTreeItem(page, constants.PROGRAM_1_NAME, "right");
+
+    await page.waitForTimeout(200);
+    await findAndClickText(page, "Copy Name");
+    await page.waitForTimeout(200);
+
+    const resNameInClipboard = await getClipboardContent(page);
+    expect(resNameInClipboard).toEqual("MYPROG1");
   });
 });