+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the IBM® CICS® Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+
+- Added headers as optional parameter to getResource and getCache methods.
+
 ## `6.11.0`
 
 - Added JVM Server CICS Resource constant